Refactor package structure (#1)
